pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "analytics-engine"
version = "0.2.0"
description = "Engagement tracking and audience analytics for ORGAN-V public-process"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4.0"]

[project.scripts]
analytics-collect = "src.goatcounter:main"
analytics-aggregate = "src.aggregator:main"
analytics-dashboard = "src.dashboard:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"src/dashboard.py" = ["E501"]
